Try to test against a page with underscore.
This doesn't work, because HtmlUnit refuses to execute JS against a plain
page?? It's not text, it's html, but maybe because it's a file??

Two alternatives - switch to Chrome, but Chrome wants some other dependency
in an environment variable.
Or switch back to loading a page that DOESN'T EXIST and the execute underscore
in HtmlUnit. Maybe that'll work.